requirements.txt
pillow
loguru